requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
boto3==1.34.104
botocore==1.34.104
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.1
django-enumfield==3.1
django-environ==0.11.2
django-sso==3.0.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jmespath==1.0.1
launchflow[gcp]
lxml==5.1.0
mysqlclient==2.2.1
openai==1.14.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-docx-replace==0.4.4
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
stripe==9.12.0
psycopg2-binary==2.9.9
